Load PLY meshes when parsing with minipbrt, for fairer comparison with pbrt-parser
pbrt-parser automatically loads PLY meshes, but minipbrt hasn't been able to before now. This meant that the performance comparison was heavily skewed in favour of minipbrt on scenes that use PLY meshes, since it was doing considerably less work.

Now that we have (work in progress) support for loading PLY meshes with minipbrt, we should use it so that we get a fairer performance comparison. That's what this commit does.