Allow a file containing a list of scenes to parse on the command line
If the user gives a filename ending in .txt as one of the (non-option) command line args, we assume it's a text file containing a list of scenes to parse. The text file must have one filename per line, with no leading or trailing whitespace.

This is hopefully a bit more convenient than writing something like `cat allscenes.txt | xargs /path/to/pbrt-parser-perf.exe`!