Include time to delete the scene in our measurements
Previously the time measurements for `pbrt-parser` included the time to delete the scene after parsing, but the time measurements for `minipbrt` and `threaded` did not. I decided it would be more illuminating to include the deletion time, so I'm now deleting inside the timed block for `minipbrt` and `threaded` as well.